Moved express-session initialisation to own file
no-issue

This is complex enough that it deserves its own module
